Add files via upload
Description of the project-
The first page consist of two options the first is choose file and second is upload.when we click on upload button then the photo gets saved on the gallery page and display page. and when we click on the view image then whole gallery page gets displayed .the name of my first page is gallery, and the second page consists of the entire gallery of photos the third page is view image ,i was supposed to make a single page and include all the contents in one page list as & have my exams i won't be able to do that .i have done all the possible things you told .i have used php and bootstrap cdn link in this project and php framework as codeigniter. i am sending the project with database .Run project using the following linkhttp://localhost/CodeIgniterUpload/                                                  .......first page
http://[::1]/CodeIgniterUpload/index.php/welcome/viewImages     ..............when click on the button of view image
i have sent project ,if you are not getting then please inform me.